+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## Unreleased
+
+## [7.2.0] - 2022-04-28
+### Added
+- Coingecko multi price fetcher
+- Multi-feeds processors sub-services for FeedProcessor
+
+### Removed
+- Coingecko single price fetcher
+
+### Changed
+- Enabled FeedProcessor integration test
+
 ## [7.1.1] - 2022-04-27
 ### Fixed
 - Updated package.json version